Message (#6)
* !Change Button Status!

* Вход по кнопке

* 2

+++ Rap/src/database/LinkConnector.java
@@ -218,7 +218,7 @@ public class LinkConnector {
 	 */
 	public static int getWorkersCount() {
 		Query q = entityManager.createQuery("SELECT COUNT(1) FROM Worker w");
-		return (Integer) q.getSingleResult();
+		return ((Long) q.getSingleResult()).intValue();
 	}
 	
 	/**
